[controller][common] Adds new controller endpoint to get node's replica readiness (part 2)
Last rb (034f48722) adds the implementation of a new endpoint in controller for getting the reaplicas state
of a server node. This rb mainly focuses on adding an integration test to cover its functionality.

It also introduces an enum NodeReplicasReadinessState to the NodeReplicasReadinessResponse class. The reason
behind is to differenciate the following two states:

1. A sever node is just added to a Venice cluster and no replicas are created on it and thenew endpoint should
   return 'ready'.
2. A sever node is stopped and the new end point should return the node is not live.

RB=2993442
BUG=VENG-8697
G=voldemort-reviewers
R=adietz,shuhliu,sdwu,xyin,smaji,mnhuang,jlliu,lcli,zpolicze,yurzhou,anaberez,fvillega,goliu,nithakka,xhao,akozma,amsaxena,xnma
A=xyin