[common][controller][da-vinci] Added controller APIs to start data recovery
Three APIs were added in order to accomondate data recovery initiated by AdminTool and fabric
buildout effectively:

1. prepareDataRecovery
2. isStoreVersionReadyForDataRecovery
3. dataRecovery

(2) and (3) should always be called for data recovery but (1) is needed when performing data recovery
on an existing fabric that might still have some lingering resources and states.

(2) is separated out because deleting those lingering resources may not be instantaneous and it's up
to the caller to decide how they would like to poll/wait for the store version to become ready to start
data recovery.

RB=3004184
BUG=VENG-8804
G=voldemort-reviewers
R=shuhliu,yurzhou,goliu
A=goliu