[Test] Fix flaky test TestWritePathComputation::testFeatureFlagMultipleDC
In testFeatureFlagMultipleDC, current test enables 'exponentialBackOff' when waiting for operation
to complete. Thus, the wait time inreases exponentially and could be larger than the timeout value
set, if it the waiting condition is not satisfied in first several rounds.

This rb changes to disable 'exponentialBackOff' for wait time.

RB=3026085
BUG=VENG-8732
G=voldemort-reviewers
R=anaberez,nithakka,mnhuang,xhao,xyin,xnma,akozma,smaji,jlliu,shuhliu,goliu,amsaxena,lcli,zpolicze,yurzhou,sdwu,fvillega
A=xhao