[test] Fix flaky test TestPushJobWithNativeReplicationt#estNativeReplicationForBatchPush
The only error I saw in the wc-test history is store not found in child DC and from the log it indicates that KMM is not fast enough to complete mirroring in 10s timeout. No other failures are detected. Increased the timeout to 30s to resolve the issue.

RB=3025828
BUG=VENG-8885
G=voldemort-reviewers
R=mnhuang,lcli,ssen,lelu,anaberez,sdwu,shuhliu,fvillega,zpolicze,xyin,adietz,xnma,smaji,akozma,nithakka,yurzhou,sumane,xhao,goliu
A=anaberez