[controller] Wait for storeConfig deletion if it is marked as deleting but points to another cluster
Problem: When a store is deleted from cluster1 and recreated in cluster2, it's possible that cluster2
child controller does not create the storeConfig because it already exists, and cluster1 child controller,
in a different thread, deletes the storeConfig, leaving no storeConfig in this child colo.

Fix: Letting cluster2 controller update or delete the lingering storeConfig will not fix the problem:
cluster1 controller could still delete the storeConfig or delete the newly created storeConfig. Hence,
instead of operating on the storeConfig, cluster2 controller just throws an exception and waits for
storeConfig deletion by cluster1.

RB=3028406
BUG=VENG-8897
G=voldemort-reviewers
R=mnhuang,yurzhou,adietz,sumane,fvillega,lelu,xyin,zpolicze,nithakka,smaji,xhao,lcli,sdwu,akozma,anaberez,goliu,ssen,jlliu,xnma
A=xyin