[controller] Make delete store idempotent for deleting Store object and StoreConfig
1. Before this change the store deletion code can only delete a store if it has both the
Store object and StoreConfig present. This is problematic because if we ever get into a
scenario where only the Store object or StoreConfig is present due to race conditions,
bugs or partial creation/deletion failure. We will pretty much be hard blocked and will
need to modify the zk directly to get out such situation.

2. The RB focuses on making delete store idempotent respect to the Store object and
StoreConfig. There could be other unknown edge cases but they are out of the scope of
this RB.

RB=3028939