[da-vinci-client] Avoids sleeping while holding lock in StoreIngestionTask::kill
StoreIngestionTask::kill has a pattern that the code is sleeping for a long period of time while
holding an object level lock. It would have performance impact if other thread tried to acquire
the same object lock and we want to avoid such scenario normally.

This rb fixes the issue by releasing the lock while it is sleeping and re-acquiring the same lock
afterwards to proceed.

Bonus: make intervalClose call close() to stop the consumer task instead of managing isRunning directly.

RB=3041426
G=voldemort-reviewers
R=goliu,akozma,jlliu,smaji,lcli,sumane,sdwu,yurzhou,xnma,xyin,ssen,nithakka,adietz,anaberez,lelu,mnhuang,zpolicze,xhao,fvillega,shuhliu
A=lcli