[test] Increase the timeout threshold of deletingHybridStoreDeletesRealtimeTopic assertions
Problem: From the log of deletingHybridStoreDeletesRealtimeTopic test, sometimes
it takes 8s (more than the threshold 5s) for a version to become online after
an empty push. As a result, the test fails when asserting the current version.

Fix: Increase the timeout threshold from 5s to 10s.

RB=3077015
BUG=VENG-8761
G=voldemort-reviewers
R=mnhuang,goliu,xhao,fvillega,xyin,akozma,zpolicze,nithakka,adietz,lelu,anaberez,sumane,lcli,ssen,smaji,yurzhou,jlliu,sdwu,xnma
A=jlliu