[h2v] ControllerClient cleanup in VPJ
There are currently 4 different ControllerClients in VPJ, not all of which
are always initialized (some are for test mocks only). This commit ensures
that the controller retry config is used uniformly by all of these clients
and that all of them get closed when VPJ gets closed. There may be further
opportunities to clean up this CC explosion, but they will be left as a
future exercise.

Also bumped VPJ's default controller retry attempt number to 5 in tests.

RB=3087928
BUG=VENG-9029
G=voldemort-reviewers
R=akozma,nithakka,sdwu,xhao,mnhuang,jlliu,yurzhou,zpolicze,xyin,anaberez,ssen,xnma,sumane,goliu,lcli,adietz,lelu,smaji,shuhliu
A=smaji