[server] Update in-memory consumedUpstreamRTOffsetMap when leader subscribes to RT
Problem: Today consumedUpstreamRTOffsetMap is only updated when a new record is consumed from RT or when server is restarted.
However, for a new push, if leader subscribes to the end of RT, no RT record will be consumed and consumedUpstreamRTOffsetMap
remains as empty. Since leader_offset_lag metric is calculated via (last RT offset - offset in consumedUpstreamRTOffsetMap),
it reports a huge lag and triggers false alarm.

Fix: When leader switches to RT, if the start offest is bigger than the offset in consumedUpstreamRTOffsetMap, update the map.

RB=3098926
BUG=VENG-8997
G=voldemort-reviewers
R=yurzhou,xyin,sumane,jlliu,lelu,adietz,goliu,xhao,sdwu,ssen,zpolicze,xnma,mnhuang,fvillega,nithakka,lcli,anaberez,akozma,smaji
A=mnhuang