[test] Disable auto.close.idle.consumers in tests
This was a tactical fix for the idle consumers issue, but it has
since been fixed more cleanly within the Kafka clients themselves,
and the dependency for this was bumped. This mode was never used
in prod, and there is no plan to use it at this time, so the tests
should be brought back in line with prod. Although the Kafka-side
fix was already validated, it is still a bit early to tell whether
it is good enough or not, so we will keep the Venice-side tactical
mitigation for now. Eventually, it could be fully removed, once we
gain more confidence with the Kafka-side fix.

RB=3103323
BUG=VENG-9056
G=voldemort-reviewers
R=adietz,sdwu,xhao,zpolicze,sumane,lelu,mnhuang,goliu,xyin,yurzhou,xnma,jlliu,akozma,lcli,ssen,nithakka,smaji,shuhliu,anaberez
A=anaberez