Added httpclient5 lib based http client in Router to talk to Venice Server
The httpclient5 lib supports both http/1.1 and http/2, and this code change
introduces the http/2 support with httpclient5 lib.
Based on the primitive benchmarking in venice-fast-client, httpclient5 based
http client could provide similar performance with http/2 enabled as httpasyncclient
with http/1.1, which is great since we have observed performance degradation
while using R2 client with HTTP/2 enabled, and this new impl could become
the formal http/2 client for all the Venice components (of coz, we will need
to benchmark/canary in prod).
This code change introduces the facility in venice-schema-common to leverage
httpclient5 to support both http/1.1 and http/2, but Venice Router will only
use it to support http/2 since http/1.1 support with httpasyncclient is already
there, and to support the full functionality of http/1.1, there would be a lot
of duplicate work, such as connection warming, so instead, Venice Router will
have to enable HTTP/2 while using this new client type.
One thing I noticed in the test suite is that httpclient2 with HTTP/2 enabled
couldn't work with JDK8, so Venice Router will fail fast if Venice Router is
using a lower version than JDK11, which should be acceptable considering all
the prod instances are JDK11 instance already.

The facility this code change included in venice-schema-common will be leveraged
by venice-fast-client later.

RB=3113921
BUG=VENG-9068
G=voldemort-reviewers
R=smaji,fvillega
A=fvillega