[Controller] Adds optional parameters to controller endpoint 'node_removable'
In the Venice InSync project, to integrate with Helix Powered Management API,
we propose to improve our existing 'node_removable' api such that it could
optionally take one extra parameter 'locked_storage_node_ids' to indicate
hosts holding a lock for maintainence. The resources on these locked hosts
would be removed at first when considering if a new node is removable. To be
compatible with today's api semantics, client (caller) could choose to not
send this parameter at all or an empty host list.

The high level design of the project is here: https://lnkd.in/geFbbYPH

RB=3175343
BUG=VENG-9092
G=voldemort-reviewers
R=xyin,goliu
A=xyin