[test] Fix leaked resources in ProcessWrapper with partial failure on internalStart
The old behavior does not cleanup properly when internalStart() fails half way. e.g. for
VeniceController we start controllerService and then adminServer. In the edge case when
adminServer cannot be started due to port conflict the ProcessWrapper does not stop the
controller completely. As a result, the half started controller will still get Helix
assignment and even become the leader of a cluster but unable to process requests and
causing tests to hang or timeout after 10 minutes...

RB=3120788
G=voldemort-reviewers
R=lelu,mnhuang,goliu,sdwu,akozma,ssen,adietz,fvillega,sumane,xnma,nithakka,smaji,zpolicze,anaberez,lcli,yurzhou,shuhliu,jlliu,xhao
A=anaberez