[compat][common][server] Updated 'upstreamOffsetMap' in 'PartitionState' to use String key
By default, Avro will use 'UTF8' as the key by default when deserializing a map,
which breaks the assumption in the code, which is using 'String' for map lookup.
To fix this issue, we add some avro properties to direct the de-serializer to use
'String' key instead instead of changing the code.
This advanced Avro feature is available from 1.7, and this field is only used in
venice backend component, so it is safe to use this Avro feature.

Ideally, we would like to deploy Venice Controller first, then Venice Server.

RB=3130421
BUG=VENG-9084
G=voldemort-reviewers
R=shuhliu,fvillega,mnhuang
A=fvillega