[fast-client] Added long-tail retry support for fast-client
So far, for single-get, Fast Client only supports dual-read, and for
some high throughput use cases, this will increase the outgoing traffic
significantly considering dual-read can be also enabled, which means
the outgoing request volume will be 3x, which could degrade the client
performance.
This RB introduces the long-tail retry feature for the single-get API,
which allows the customer to configure the retry threshold.
Also if the original request fails, the retry will be triggered right
away to improve the availability.

Currently, this change doesn't limit the total amount of retry requests,
and we could include this as a later improvement, and also we may want
to introduce the smart long-tail retry logic to fast-client as well to
avoid useless retries. Lastly, we need to add long-tail retry support
for batch-get APIs as well.

RB=3135246
BUG=VENG-9037
G=voldemort-reviewers
R=mnhuang,akozma,jlliu,sdwu,anaberez,lcli,sumane,fvillega,adietz,yurzhou,ssen,lelu,xyin,nithakka,smaji,shuhliu,xnma,xhao,zpolicze
A=smaji