[test] Fix a regression reported by SpotBug
SpotBug reported below error:
This method accesses the value of a Map entry, using a key that was
retrieved from a keySet iterator. It is more efficient to use an
iterator on the entrySet of the map, to avoid the Map.get(key) lookup.

Fix is to iterate on the map values.

RB=3133340
G=voldemort-reviewers
R=goliu,lelu,sdwu,xnma,yurzhou,nithakka,mnhuang,zpolicze,fvillega,xhao,sumane,smaji,akozma,shuhliu,xyin,anaberez,adietz,jlliu,ssen
A=goliu