[server][da-vinci] Fix Active/Active ingestion metrics issue in forked ingestion process
A/A metrics are not displayed in forked ingestion process as the forked process does not parse the local region name correctly. From server log it shows correct region name in main process but null in forked process.
This RB fixes the issue by setting the region config when building forked process config.

RB=3133681
BUG=VENG-9072
G=voldemort-reviewers
R=xyin,fvillega,ssen,smaji,sdwu,yurzhou,sumane,adietz,nithakka,akozma,shuhliu,mnhuang,zpolicze,xhao,lelu,anaberez,lcli,xnma,goliu
A=smaji