[server] Building quota enforcer after partition count is prepared.
If we build quota enforcer too early in StoreIngestionTask constructer, the subPartitionCount is not ready, then it will fail the integration test.
Previous we did not find it from CI test, because the CI test keeps retrying on failed test cases until timeout the build.

RB=3135774
BUG=VENG-8995
G=voldemort-reviewers
R=fvillega,yurzhou,sdwu,xyin,zpolicze,lelu,nithakka,smaji,ssen,anaberez,jlliu,xnma,adietz,mnhuang,akozma,shuhliu,sumane,goliu,lcli
A=mnhuang