[controller] Minor KafkaAdmin and test improvements
Main code changes:

1. The KafkaAdmin's getAllTopicConfig is replaced by getSomeTopicConfigs
which gets passed a list of topic names. The only usage of this function
already knows what topics it's interested in, and this refactoring
eliminates one call to listTopics().

2. Removed boilerplate in the InstrumentedKafkaAdmin and instrumented
the close function as well.

Test changes:

1. Added various debugging enhancements for flaky tests.

2. Also deleted ProducerConsumerReaderIntegrationTest since this path is
highly exercised by a bunch of other tests, and it was a bit messy to
begin with, so not worth cleaning up.

3. Fixed a regression in waitForNonDeterministicAssertion which made
the exponential backoff param irrelevant.

RB=3157935
BUG=VENG-9139
G=voldemort-reviewers
R=akozma,ssen,nithakka,xhao,smaji,sdwu,yurzhou,jlliu,xyin,lelu,xnma,sumane,goliu,zpolicze,lcli,anaberez,mnhuang,shuhliu,adietz
A=mnhuang