[venice-push-job] Add config to suppress transmitting EOP
This is a waterloo ask.  Essentially they still go through the flow where they want to be able to ETL the future version, however currently the venice push job always transmits the EOP message, meaning that once the push is complete, it'll begin serving traffic.  With this change, waterloo users will be able to do what they do currently by selecting in Nuage that the push was successful or should be aborted after analyzing ETL results of the future version.

This change is going to seem strange as you might ask why not analyze the data in the offline.  This change is meant to go hand in hand with the notion that the future version will have RT data replayed on top of it.  This is accomplished by setting the offsetLag catchup to -1 and letting the rewind run continously until the config is set post analysis.

RB=3173713
G=voldemort-reviewers
R=sdwu,lelu,lcli,mnhuang,jlliu,adietz,sumane,akozma,smaji,xhao,xyin,ssen,xnma,anaberez,goliu,nithakka,yurzhou,fvillega,shuhliu
A=nithakka