[da-vinci][server][controller] Refactored ingestion task constructors
The constructor of StoreIngestionTask and all of its sub-classes were
really ugly, with tons of parameters. Worse still, most of the params
being passed to the constructor of the sub-classes were just passed
through to the super constructor, without the subclasses looking at
them.

The new structure lets the builder object coming from the factory
flow down through all of the ingestion task classes, thus greatly
simplifying all of the constructors.

It's possible that there may be even more opportunities to make clean
these constructors, but for now only the params available in the
builder are removed, in an effort to minimize this change's risk of
regression.

Unrelated:

- Added a null check in VeniceHelixAdmin#getOffLinePushStatus() so
  that it throws a VeniceNoStoreException rather than a NPE. This
  should not increase the number of code paths calling this which
  can fail, but merely make them fail with a more sensible error
  message.

RB=3180879
BUG=VENG-9179
G=voldemort-reviewers
R=ssen,xyin,akozma,smaji,mnhuang,lcli,goliu,xnma,xhao,nithakka,jlliu,sumane,yurzhou,zpolicze,lelu,anaberez,adietz,shuhliu,sdwu
A=smaji