[Test] Uses a dedicated ZkServer for each test fixture
Today, we have long a running singleton Zk server thread shared by all test fixtures and that leads to
occastional timeout issues in CI tests. Through a number of experiments, we have observed that after
creating dedicated ZkServers for each test fixture instead of sharing a singleton across all fixtures,
the timeout issues have disappeared. We have ideas about spinning up ZK in a forked process for better
resource isolation, stability, and many other benefits.

So this RB could be a temporary solution before we are confident about the change to make ZK a forked
process. It introduces a swith that allow us to use dedicated ZkServers for each test fixture. ZkServer
threads are closed when a test fixture is completed.

This rb, in fact, does not add many new logics, it reverts back some related codes in a previous
RB where we removed the SINGLETON in ZkServerWrapper to address the spotbug errors.

https://rb.corp.linkedin.com/r/3047098/

RB=3187536
G=voldemort-reviewers
R=fvillega,anaberez
A=xhao