[server] Make sure forked ingestion process cleans up lingering metadata partitions in server
Previously when ingestion isolation is enabled in server, the metadata partition will never be deleted as the offline->dropped action only triggers data partition being dropped without dropping metadata partition opened in the forked process. This results in more than ten thousands of store versions lingering in the server hosts and impact the start up time heavily.

Da Vinci is not impacted as it has the version deletion logic in the version swap and bootstrap also deletes stale versions.

This RB provides the fix to delete the metadata partition when there is no data partition open in both processes.

RB=3204128
G=voldemort-reviewers
R=lcli,anaberez,xyin,xhao,yurzhou,adietz,nithakka,lelu,goliu,shuhliu,smaji,zpolicze,sdwu,xnma,akozma,mnhuang,fvillega,sumane,ssen
A=nithakka