[thin-client] Introduced a new mode to force cluster discovery at startup time
This code change introduces a new option in client config:
forceClusterDiscoveryAtStartTime
When this config is specified, the startup method of the Venice Client
will do initialization and if the initialization fails, the client will
propagate the exception to the application.

This config would only work properly if d2 client is already started.
With this option, Venice Client will force the cluster discovery at
startup time and avoid cluster discovery at runtime, which could be
problematic if the cluster discovery request is problematic, which could
block the Venice request for a fairly long time, which could occupy all the
application threads.

This code change also disables the cluster discovery retry at runtime, and
we will rely on the following Venice requests to retry the cluster discovery
if it fails.

RB=3207523
BUG=VENG-9205
R=lelu,sdwu,smaji,zpolicze,akozma,adietz,nithakka,lcli,ssen,jlliu,xhao,fvillega,xnma,yurzhou,xyin,anaberez,sumane,mnhuang,shuhliu
A=xyin