[test] Sets org.apache.zookeeper package log level to error
The tests generate large amount of logs when running in CI and it could cause
the Jenkins framework to run into the out of memory issues. Given that a good
amount of logs are generated by the ZK server which Venice does not care very
much, we can reduce the log size by setting higher log level to ZK server.

This rb sets log level to error for org.apache.zookeeper packages.

RB=3216091
G=voldemort-reviewers
R=anaberez,ssen,jlliu,smaji,adietz,lelu,akozma,xyin,yurzhou,goliu,sumane,nithakka,zpolicze,mnhuang,sdwu,lcli,xnma,xhao,fvillega,shuhliu
A=anaberez