[da-vinci][server] Minor optimization to shared consumer logic
The sanitizeTopicsWithoutCorrespondingIngestionTask function in the
SharedKafkaConsumer used to be O(N) to the number of messages consumed,
and with this change it is now O(N) to the number of unique partitions
included in the current batch of messages.

Changed the return type of getIngestionTaskForTopicPartition so that
it returns a StoreIngestionTask rather than an Optional<SIT>.

Added a benchmark for sanitizeTopicsWithoutCorrespondingIngestionTask
with permutations for:
- Topic-wise and partition-wise implementations
- The number of messages in the batch (100, 1K, 10K, 100K)
- The number of partitions in the batch (1, 10, 100)

Based on this benchmark, the throughput is greatly improved and GC is
reduced. For example, for throughput:
- The partition-wise/100/1 goes from 0.8 to 40 op/microsecond.
- The partition-wise/10K/100 goes from 0.008 to 0.9 op/microsecond.
- The topic-wise/100/1 goes from 0.9 to 35 op/microsecond.
- The topic-wise/10K/100 goes from 0.008 to 1 op/microsecond.

Allocation rate was anywhere between 88 B and 3.2 KB per op, and is
now a flat 56 B for all permutations.

The build as a whole seems to be faster as well, going from an
average of 52:42 over the last 10 commits down to 48:40 with this
commit (sample size = 6).

The unit tests are refactored to attempt to capture similarities
between the partition-wise and topic-wise implementations, and a new
test is added specifically for the optimized function.

RB=3221906
BUG=VENG-9251
G=voldemort-reviewers
R=mnhuang,zpolicze,smaji,akozma,sdwu,yurzhou,jlliu,sumane,xhao,lcli,adietz,nithakka,lelu,shuhliu,anaberez,xyin,ssen,goliu,xnma
A=smaji