[controller] Graceful shutdown for admin message generation
1. Acquire the cluster resource read lock so during graceful shutdown we will
not generate an admin message and bump up the execution id but not sending it
to the admin topic which creates a gap that blocks consumption.

2. The change avoids execution id gap by acquiring a cluster read lock before
generating and sending the message. During this time the shutdown cannot acquire
the write lock which means all previously generated admin messages will be sent
before shutdown can kick off. Once shutdown is in progress future requests to
generate new admin messages should be blocked because the write lock is acquired.

RB=3229939
BUG=VENG-9210
G=voldemort-reviewers
R=ssen,lelu,zpolicze,shuhliu,fvillega,anaberez,sdwu,goliu,yurzhou,jlliu,xnma,sumane,xhao,smaji,adietz,mnhuang,akozma,lcli,nithakka
A=lcli