[AA][controller] Make parent controller not update existing RMD schema with the same value schema ID and RMD version ID
This change is to fix a bug where the parent controller generated an RMD
schema and tried to use this new RMD schema to overwrite an existing RMD
with the same value schema ID and RMD version ID. This attempt to
overwrite is supposed to fail because such schema overwrite is not
supprted. This change avoids trying to overwrite existing RMD schemas.

RB=3232894
BUG=VENG-9262
G=voldemort-reviewers
R=yurzhou,xnma,sumane,smaji,anaberez,nithakka,shuhliu,xyin,adietz,zpolicze,akozma,sdwu,mnhuang,fvillega,lelu,goliu,ssen,jlliu,xhao
A=zpolicze,goliu