[server][controller][h2v][compat] Add version swap deferred flag as option to add version api and VPJ and store version config
1. child controller must be deployed before parent controller
2. controllers must be deployed before servers.

This is to facilitate the ask from waterloo for the ability to start RT consumption on a hybrid store but to NOT swap the current and future versions.  The way this works is that we add a flag to a store version config which determines if we need to defer version swap which is honored by the controller.  A verison pushed this way needs to be marked manually with the 'makeCurrentVersion' admin api in order to be a serving version.  If this version is not marked to serve and another push starts, it will be cycled out.

RB=3237426
G=voldemort-reviewers
R=ssen,xyin,yurzhou,sdwu,anaberez,nithakka,shuhliu,smaji,sumane,xhao,akozma,lcli,jlliu,fvillega,lelu,adietz,xnma,mnhuang,goliu
A=nithakka