[da-vinci][server][router] GC optimization in KafkaConsumerService run loop
Eliminated several instantiations from the run loop of shared consumers,
including several per-poll instantiations and, more importantly, some
per-message instantiations. This is achieved by removing some unneeded
massaging of the poll results, which can be accomplished by an API that
is already available out of the box on ConsumerRecords. Unfortunately,
there is still a O(N) operation proportional to the number of polled
messages, which cannot be entirely eliminated if bandwidth throttling
is enabled. In this patch, however, we do at least eliminate this O(N)
operation in cases where bandwitdh throttling is disabled.

Unrelated:

1. Changed all usages of Supplier<some primitive type> to the classes
   dedicated for this by the JDK (e.g. LongSupplier) to avoid boxing.

2. Moved some stats-related classes from common to da-vinci-client
   since they were only used over there.

RB=3246292
BUG=VENG-9293
G=voldemort-reviewers
R=xhao,xnma,sdwu,mnhuang,xyin,sumane,zpolicze,nithakka,lcli,lelu,ssen,shuhliu,smaji,akozma,adietz,goliu,yurzhou,anaberez,jlliu
A=xhao