[h2v] Remove delete record in push status system store logic after incremental push job
This method to be removed may trigger ACL issue if the actual job runner is not in part of user store read ACL list, for example some proxy headless accounts.
If we really need to remove old records, we will cut another RB to add logic to use controller client to delete the record, as controller will always have the permission to write to topics.

RB=3247625
G=voldemort-reviewers
R=lelu,anaberez,akozma,sumane,shuhliu,nithakka,adietz,smaji,xyin,xnma,sdwu,lcli,ssen,fvillega,xhao,yurzhou,mnhuang,zpolicze,goliu
A=sumane