[venice-common] Fix a SpotBugs warning and a potential race in HelixHybridStoreQuotaRepository
In onHybridStoreQuotaViewChange, SpotBugs gives "ML_SYNC_ON_FIELD_TO_GUARD_CHANGING_THAT_FIELD" warning and
complains that it synchronizes on the field 'resourceToStatusMap' and it also tries to update that field,
which might break the mutual exclusion gurantee.

Meanwhile, for all the places that resourceToStatusMap needs to be read, it should also acquire a lock
(same as before) to avoid potential races.

This rb fixes the issue by introducing a lock to protect resourceToStatusMap and also other places
where resourceToStatusMap is read.

RB=3249726
G=voldemort-reviewers
R=fvillega,adietz,smaji,lcli,xnma,anaberez,xyin,jlliu,zpolicze,shuhliu,nithakka,goliu,mnhuang,akozma,xhao
A=xhao