[test] Fix flaky test testZstdCompressingVsonRecord
In testZstdCompressingVsonRecord, the test code sleeps 1 sec before query router for the stored data
in order to compensate for the time that DictionaryRetrievalService requires to download the dictionary.
One sec sleeping time could be short and we should use waitForNonDeterministicAssertion for such aync
operation to complete.

One logline in DictionaryRetrievalService is moved to a later part, so that we know exactly if related
compressor is ready.

Bonus: fix similar issues in TestBatch.

RB=3268221
G=voldemort-reviewers
A=nithakka