[da-vinci] Avoid duplicate metric issue caused by re-initializing StoreBackend object
This issue is triggered by the following logic:
When a store is removed from Da Vinci application's managed config list, our bootstrap logic will try to remove all associated versions ingested locally.

However, the order is incorrect here: when we itertaing the store engine repo, we will try to create StoreBackend if not exists. For unmanaged store, we will (1) first create StoreBackend(which register StoreBackendStats) and (2) delete StoreBackend(however, it won't unregister metrics) and (3) remove storage engine. The problem is, when we see a different version of the same store, we will try to do (1) again, which will trigger the above mentioned exception.

The proposed change adjust the flow: Instead of deleting the StoreBackend immediately, it will collect them together and delete them after scanning all storage engines.

I've also consider changed the metric register flow, but this won't work for now. If we have a need to recreate the StoreBackend in the future, that logic might lead to fail to locate the sensor, as we do not have existing API to retrieve existing sensor by metric name.

RB=3257299
BUG=VENG-9301
G=voldemort-reviewers
R=xnma,sdwu,xhao,nithakka,anaberez,zpolicze,adietz,ssen,xyin,sumane,lcli,akozma,goliu,smaji,lelu,fvillega,shuhliu,yurzhou,mnhuang
A=anaberez