[da-vinci][server] Removed record wrapper and other GC optimizations
The ingestion code previously did a trivial wrapping of all consumer records
such that all record wrappers coming from a given consumer instance and
belonging to the same topic-partition always contained the same metadata.
This was a O(N) operation proportional to the number of messages consumed.

The main change in this commit is to remove this wrapping and instead add
the metadata which the wrappers redundantly carried through all of the
functions that need it.

Other changes:

- Many calls to compute the subPartition are now eliminated.
- Added a dependency on fast util, and started using some of its sets
  and maps with primitive types.

RB=3257705
BUG=VENG-9321
G=voldemort-reviewers
R=anaberez,mnhuang,nithakka,zpolicze,akozma,adietz,jlliu,lcli,xnma,lelu,sdwu,shuhliu,xhao,smaji,goliu,sumane,yurzhou,ssen,xyin
A=smaji