[controller] Change how superset schema generation logic sets default value on the same field in the generated superset schema.
The original behavior is that default value from the old schema is used
if a field exists in both the old and new schemas. However, we decided
to use the default value from the new schema instead.

Other changes are to rename local variable names to be precise about
which schema is the old schema and which schema is the new schema.

RB=3265787
G=voldemort-reviewers
R=xnma,yurzhou,xyin,ssen,sumane,nithakka,shuhliu,goliu,sdwu,akozma,xhao,anaberez,mnhuang,fvillega,smaji,adietz,lelu,zpolicze,jlliu
A=smaji