[controller] Remove getReplicationMetadataVersionId method from the Admin (controller) interface and add a method named getReplicationMetadataSchema to the Admin interface
getReplicationMetadataVersionId is a method used to check whether
generated RMD schema has been successfully added or not. However this
function is flawed because it takes in one RMD schema string and returns
only one RMD version whereas one RMD schema string can have multiple RMD
versions. Instead of changing the signature of this
getReplicationMetadataVersionId method to let it return a list of RMD
versions, add another method named getReplicationMetadataSchema which
serves the same purpose but in a simpler way.

RB=3268844