[venice-common] Fix a SpotBugs warning (race condition) in RoutingDataRepository
In HelixExternalViewRepository::onExternalViewDataChange, SpotBugs reports
"ML_SYNC_ON_FIELD_TO_GUARD_CHANGING_THAT_FIELD" warning because it synchronizes on the field
'liveInstancesMap' and at the same tries to update it, which could lead to synchronize on different
objects and not wait until the synchronized block finishes. It is not a recommended code pattern.

Meanwhile, for all the places that liveInstancesMap is read, it should also acquire the same lock
to avoid potential races.

This rb fixes the issue by introducing a lock to protect liveInstancesMap and making sure any
liveInstancesMap reads acquire the lock.

RB=3279256
G=voldemort-reviewers
A=anaberez