[Server] Release the KafkaProduer buffer when the ingestion task is closed
In Prod, we observed some OOM issue, which is caused by leaked StoreIngestionTask,
which will leak the buffers allocated by the internal KafkaProducer.
This code change will reset/release the internal KafkaProducer as long as
the ingestion task is closed, and in the meantime, this code change added
a new heartbeat metric for ParticipantStoreConsumptionTask thread, so
we could setup alert against this metric.

RB=3287106
BUG=VENG-9285
R=jlliu,mnhuang,nithakka,zpolicze,yurzhou,sdwu,smaji,shuhliu,anaberez,adietz,xhao,fvillega,akozma,ssen,lcli,xyin,sumane,xnma,lelu
A=fvillega