[server] Add a "owner" field to TopicPartitionWithURL without functional change.
This "owner" field is needed to support a situation where a topic partition
in the same Kafka cluster needs to be consumed by multiple threads
concurrently on the same host. For example, 2 StoreIngestionTask instances
on the same Venice server host, both of them consume from the same partition
in the real-time topic. With this "owner" field, these 2
StoreIngestionTask instances could provide their own "owner" string to
differentiate the 2 TopicPartitionWithURL instances.

RB=3289271
G=voldemort-reviewers
R=smaji,jlliu,sdwu,goliu,xyin,anaberez,zpolicze,fvillega,shuhliu,sumane,ssen,mnhuang,nithakka,xhao,akozma,xnma,lelu,adietz,yurzhou
A=xhao