[controller][da-vinci][server] AutoCloseableLock efficiency optimization
The AutoCloseableLock is refactored in order to allocate fewer objects.

AutoCloseableLock is made abstract, and includes the safeguard against
multiple invocations of close(). This is refactored to use a primitive
boolean rather than an AtomicBoolean, for reasons explained in the
JavaDoc of the function.

AutoCloseableLock also provides static factory methods for the cases
of wrapping a single lock and multiple locks. These return instances
of new classes: AutoCloseableSingleLock and AutoCloseableMultiLock,
both of which minimize object allocations for their respective use
cases.

TestClusterLockManager is improved to ensure that locks are properly
released and reusable. TestResourceAutoClosableLockManager is also
added to further improve coverage.

RB=3301965
BUG=VENG-9511
G=voldemort-reviewers
R=jlliu,goliu,nithakka,smaji,xhao,xnma,xyin,anaberez,sumane,shuhliu,lelu,sdwu,akozma,mnhuang,ssen,adietz,yurzhou,zpolicze,lcli
A=lcli